Rename SanitiseConstructor case to SanitiseDeleteConstructor.
