+++ buildSrc/jarfilter/src/test/resources/sanitise-delete-constructor/build.gradle
@@ -8,7 +8,7 @@ sourceSets {
     main {
         kotlin {
             srcDir files(
-                '../resources/test/sanitise-constructor/kotlin',
+                '../resources/test/sanitise-delete-constructor/kotlin',
                 '../resources/test/annotations/kotlin'
             )
         }
@@ -21,7 +21,7 @@ dependencies {
 }
 
 jar {
-    baseName = 'sanitise-constructor'
+    baseName = 'sanitise-delete-constructor'
 }
 
 import net.corda.gradle.jarfilter.JarFilterTask
